consensus: avoid using Wrap in WALToProto
Construct the inner Message directly using the wrapper variant types
defined in the generated proto bindings.
Amend the general contract of the Wrapper interface to not
expect the returned protobuf message be of a particular type.
The reason is, we implement Wrap with the Message type of the same
package where the variant field type is defined (we can't import
future versions' packages as they contain references to types in
older versions, which would result in an import cycle),
so it may end up not being the latest version of Message.
For the wire format it's all the same, as long as the Message encoding
remains backward-compatible.